Created doc_str! macro. About 95% complete. Works almost exactly like format!(). Honestly, it's basically just a wrapper around format!
The purpose of this was to create a macro for easily making
a multilined, single #[doc] attribute for compiled
structs/enums/functions.
Hoenstly, I'm not sure if I can even use this in the way I was hoping
to. Since you can't use a proc_macro from within the same crate that it
is defined in. Unless if I extract doc_str into it's own sub-crate..