Reverted back to original Attribute implementation for now. Attribute now handles a variety of Restify Attribute Commands.   - Rename   - ReanmeAll   - Buidler   - SkipIf   - Default   - Derive Adding more attribute commands in the future will also be easier to do.
There is one glaring issue that I still need to figure out though.
Detecting if a user if trying to add a Type-Specific attribute to
a Type Parameter, and vise-versa.
Example: Restify should panic if it finds a #[derive(...)] attached to a Type parameter.

Before reverting back to this version of Attribute. Detecting and
managing this exact sitatino is what I was trying to implement, without
having to resort to two seperate implementaions of Attribute. I tried to
implement it with both generic traits and with the Type State Pattern.
Both attemps ended up becoming convoluted and unmanageable