Realized that I needed to implement logic for Restify's Generator to determin if an Attribute is either a Quotable Attribute or a Command Attribute.
Changes I made to make this possible, without having to completely
rewrite everything( and go through the entire project, and update
everything)

First, I created a few new data types.
 - enum AttributeCommands:
    * Will hold Command Identifiers for Restify to use when it's
      getting ready to generate.
 - enum AttributeType {
     Quote(proc_macro2::TokenStream),
     Command(AttributeCommands),
   }
 - pub struct CompiledAttributes {
    pub quotes: Vec<proc_macro2::TokenStream>,
    pub commands: Vec<AttributeCommands>
   }

Before any of the Generation steps. Restify will first take any of the
Attribute types that's provided(Attributes, AttributeSlice) and make a
call to the correct methods that will Convert them into a CompiledAttributes
object.

When a method's parameter type is AttributeSlice<A>,
  let compiled_attributes: CompiledAttributes = attributes.into();
When a method's parameter type is &Attributes<A>
  let compiled_attributes = &field.attributes.compile();

Now, Restify's Generator will be able to take quotes, and include them
within the generated code. And use commands for specialized pre/post
generative steps.