After implementing my original Command trait. I came to the conclusion that it would not work out as I'd hoped. Working with Command implemented types would require Dynamic Displatch, for this use-case.
And the trait Command I originally created is not Object safe.

I stashed the trait definition and BuildCmd implementation in
commmand/failed_command.rs. Only as a potential reference for in the
future.

For now though.. I'm going to resort to simply calling outside methods
based on which enumeration AttrCommands is found to be.